move files from util to simulator
