+++ chia/simulator/simulator_constants.py
@@ -1,7 +1,7 @@
 if __name__ == "__main__":
     from chia.util.default_root import DEFAULT_ROOT_PATH
     from chia.simulator.block_tools import create_block_tools, test_constants
-    from chia.util.keyring import TempKeyring
+    from chia.simulator.keyring import TempKeyring
 
     with TempKeyring() as keychain:
         # TODO: mariano: fix this with new consensus
