+++ tests/core/data_layer/test_data_rpc.py
@@ -28,7 +28,7 @@ from chia.types.peer_info import PeerInfo
 from chia.util.byte_types import hexstr_to_bytes
 from chia.util.config import save_config
 from chia.util.ints import uint16, uint32
-from chia.util.setup_nodes import SimulatorsAndWalletsServices
+from chia.simulator.setup_nodes import SimulatorsAndWalletsServices
 from chia.wallet.trading.offer import Offer as TradingOffer
 from chia.wallet.transaction_record import TransactionRecord
 from chia.wallet.wallet import Wallet
