+++ tests/simulation/test_simulation.py
@@ -16,11 +16,11 @@ from chia.simulator.simulator_protocol import FarmNewBlockProtocol, GetAllCoinsP
 from chia.simulator.time_out_assert import time_out_assert
 from chia.types.peer_info import PeerInfo
 from chia.util.ints import uint16, uint32, uint64
-from chia.util.setup_nodes import SimulatorsAndWallets, setup_simulators_and_wallets, setup_full_system
-from chia.util.setup_services import setup_full_node
+from chia.simulator.setup_nodes import SimulatorsAndWallets, setup_simulators_and_wallets, setup_full_system
+from chia.simulator.setup_services import setup_full_node
 from chia.wallet.wallet_node import WalletNode
 from chia.simulator.block_tools import test_constants
-from chia.util.keyring import TempKeyring
+from chia.simulator.keyring import TempKeyring
 
 test_constants_modified = test_constants.replace(
     **{
