+++ tests/wallet/rpc/test_dl_wallet_rpc.py
@@ -13,7 +13,7 @@ from chia.simulator.time_out_assert import time_out_assert
 from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.types.peer_info import PeerInfo
 from chia.util.ints import uint16, uint32, uint64
-from chia.util.setup_nodes import SimulatorsAndWalletsServices
+from chia.simulator.setup_nodes import SimulatorsAndWalletsServices
 from chia.wallet.db_wallet.db_wallet_puzzles import create_mirror_puzzle
 from tests.util.rpc import validate_get_routes
