+++ tools/generate_chain.py
@@ -17,7 +17,7 @@ from chia.types.blockchain_format.coin import Coin
 from chia.types.spend_bundle import SpendBundle
 from chia.util.chia_logging import initialize_logging
 from chia.util.ints import uint32, uint64
-from chia.util.keyring import TempKeyring
+from chia.simulator.keyring import TempKeyring
 from tools.test_constants import test_constants
